fix(agent-skill-engineering): move this slice's evidence guards out of the pack tree
`tools/lint-pack-test-boundary.py`'s pack-tests-stay-in-pack check reported 32
findings against this slice. Pack tests may not reach above their owning pack,
and this slice's guards did, in two ways.

Repository-level reads: SPEC_DIR, TRANSCRIPT_ROOT and _at_base's cwd at the
repository root. The frozen spec, its verification ledger and the retained
graded transcripts are slice evidence under docs/specs/, not pack content, and
`git show <base>` runs from the repository root.

Joins a static reading cannot confine: AUTHOR_ROOT / declared, where declared
comes from evals.json at runtime. The pack module states the rule it broke on
its own line 16 — paths are literal "so every path this suite opens is
statically confined to its own pack". Checking containment at runtime is a
repository-level guarantee, not a pack-local one.

Eight guards, two helpers and the retained-transcript host-identity scan move
to tests/roster/test_ase_composition_fixture_evidence.py, which is where the
lint's own message and an existing sibling roster module both put this. The
lint has no allowlist for the check, so the destination was forced rather than
chosen. The split is by what a guard has to read, not by which criterion it
serves: several criteria are now covered from both sides, and every property
the plan named still has a guard.

Nothing is restated across the boundary. The host-identity pattern strings are
loaded from the pack suite under a unique module name, so the two trees cannot
drift to different pattern lists, and an emptied set reddens rather than
passing vacuously. The admitted eval-id set is derived from the declarations
for the same reason, rather than copying the pack suite's pinned enumeration.

Three mutations re-proved after the move: a seeded host path in a transcript, a
symlink out of the transcript root, and an emptied pattern set at its source.

plan.md and spec.md name the old module and stay frozen; the divergence is
recorded in the verification ledger, along with why eight review rounds missed
this. The gate was never run — the local gate set was an inherited list that
did not include make build-check, and it surfaced on the first CI run. An
inherited gate list is not the gate set.

Gates: pack 236; the new roster module 9; pack-test-boundary, spec-status,
brief-coverage, catalogue lint --deep, catalogue verify, ci-parity all exit 0;
diff --check clean.
